good enough!
The card parsing is half-assed... we don't do power or toughness or loyalty, since they aren't classed on mtg.wtf (making them less trivial to isolate). And we haven't worried at all about blocking out the card name in the plot. But that shouldn't be too bad... it's just a matter of adding rectangles to an existing Matplotlib window, right?